backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.3.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis & Celery
redis>=5.0.1
celery>=5.3.6

# Milvus
pymilvus>=2.3.6

# AWS/S3
boto3>=1.34.34

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# OCR & Document Processing
docling>=1.0.0
pillow>=10.2.0
opencv-python>=4.9.0.80
pytesseract>=0.3.10

# LLM Providers
openai>=1.10.0
anthropic>=0.8.1
mistralai>=0.0.12

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.3

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
